# OpenAPI/coffeeshops.yaml
openapi: 3.1.1
info:
  title: CoffeeShops
  description: REST API для системы управления кофейнями, напитками и пользователями
  version: 1.0.0
  contact: {}
servers:
  - url: /api/v1
    description: Основной сервер API
    
#-----------------------------------------SCHEMAS-----------------------------------------------
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "Описание ошибки"
...
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Уникальный идентификатор категории
          example: "9b165097-1c9f-4ea3-bef0-e505baa4ff63"
        name:
          type: string
          description: Название категории
...
    Drink:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Уникальный идентификатор напитка
        name:
          type: string
          description: Название напитка
        # categoryIds:
        #   type: array
        #   items:
        #     type: string
        #     format: uuid
        #   description: Список идентификаторов категорий, связанных с напитком
...
    CoffeeShop:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Идентификатор кофейни
        companyId:
          type: string
          format: uuid
          description: Идентификатор компании
        address:
          type: string
          description: Адрес кофейни
        workingHours:
          type: string
          description: Часы работы кофейни
      example:
        id: "123e4567-e89b-12d3-a456-426614174000"
        companyId: "123e4567-e89b-12d3-a456-426614174001"
        address: "Москва, Тверская, 10"
        workingHours: "08:00-22:00"
...
  parameters:
...
    DrinkId:
      name: drinkId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Произошла внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
...
  /drinks/{drinkId}:
    get:
      tags:
          - Drinks
      summary: Получить напиток по ID
      description: Возвращает информацию о напитке по его идентификатору
      operationId: GetDrinkById
      parameters:
        - $ref: '#/components/parameters/DrinkId'
      responses:
        '200':
          description: Информация о напитке
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Drink'
        '404':
          $ref: '#/components/responses/NotFound'

        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Drinks
      summary: Удалить напиток
      description: Удаляет напиток по дентификатору (только для модераторов и администраторов)
      operationId: DeleteDrink
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DrinkId'
      responses:
        '200':
          description: Напиток успешно удален
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /drinks/{drinkId}/categories:
    get:
      tags:
        - Drinks
      summary: Получить категории напитка
      description: Получает все категории, связанные с конкретным напитком
      operationId: GetDrinkCategories
      parameters:
        - $ref: '#/components/parameters/DrinkId'
      responses:
        '200':
          description: Категории напитка успешно получены
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /coffeeshops/{CoffeeShopId}:
    delete:
      tags:
        - CoffeeShops
      summary: Удалить кофейню
      description: Удаляет кофейню по дентификатору (только для модераторов и администраторов)
      operationId: DeleteCoffeeShop
      security:
        - bearerAuth: []
      parameters:
        - name: CoffeeShopId
          in: path
          required: true
          schema:
            type: string
          description: Идентификатор кофейни
      responses:
        '200':
          description: Кофейня успешно удалена
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      tags:
          - CoffeeShops
      summary: Получить кофейню по ID
      description: Возвращает информацию о кофейне по ее идентификатору
      operationId: GetCoffeeShopById
      parameters:
      - name: CoffeeShopId
        in: path
        required: true
        schema:
          type: string
        description: Идентификатор кофейни
      responses:
        '200':
          description: Информация о кофейне
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoffeeShop'
        '404':
          $ref: '#/components/responses/NotFound'

        '500':
          $ref: '#/components/responses/InternalServerError'

  #---------------------------------FAVORITE COFFEE SHOPS----------------------------------------------------
...